j: Add the jqt ide
Builds the jqt ide and includes it in the j derivation if the withJQT
option is true.

Can be tested in a nix-shell like so:
$ nix-shell -I nixpkgs=/path/to/nixpkgs \
 -p 'j.override { withJQT=true; }' --run jqt

Or on wayland:
$ nix-shell -I nixpkgs=/path/to/nixpkgs \
 -p 'j.override { withJQT=true; }' \
 -p qt5.qtwayland --run jqt

Work is still in progress on full functionality in jqt.  The main window
launches but not much else seems to be working.